lulu_executive_dashboard/lulu_executive_dashboard/requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
scikit-learn>=1.4.0
python-dateutil>=2.9.0
